Refactor code into smaller packages
