+++ cmd/keys.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/unweave/cli/config"
+	"github.com/unweave/cli/ssh"
 	"github.com/unweave/cli/ui"
 	"github.com/unweave/unweave/api/types"
 	"github.com/unweave/unweave/tools"
@@ -41,7 +42,7 @@ func sshKeyAdd(ctx context.Context, publicKeyPath string, owner string, name *st
 		name = &filename
 	}
 
-	uwc := InitUnweaveClient()
+	uwc := config.InitUnweaveClient()
 	params := types.SSHKeyAddParams{
 		Name:      name,
 		PublicKey: string(publicKey),
@@ -67,61 +68,13 @@ func SSHKeyAdd(cmd *cobra.Command, args []string) error {
 	if len(args) == 2 {
 		name = args[1]
 	}
-	keyname, err := sshKeyAdd(cmd.Context(), publicKeyPath, config.Config.Unweave.User.ID, &name)
+	keyname, err := ssh.Add(cmd.Context(), publicKeyPath, config.Config.Unweave.User.ID, &name)
 	if err != nil {
 		return err
 	}
 
 	ui.Successf("SSH key added as %q", keyname)
 	return nil
-}
-
-func getSSHKeyFolder() string {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		ui.Errorf("Unable to find home directory: %s", err)
-		os.Exit(1)
-	}
-	dotSSHPath := filepath.Join(home, ".ssh")
-	if _, err := os.Stat(dotSSHPath); os.IsNotExist(err) {
-		if err := os.MkdirAll(dotSSHPath, 0700); err != nil {
-			ui.Errorf(".ssh directory not found and attempt to create it failed: %s", err)
-			os.Exit(1)
-		}
-	}
-	return dotSSHPath
-}
-
-func sshKeyGenerate(ctx context.Context, owner string, name *string) (keyName string, pub []byte, err error) {
-	uwc := InitUnweaveClient()
-	params := types.SSHKeyGenerateParams{Name: name}
-	res, err := uwc.SSHKey.Generate(ctx, owner, params)
-	if err != nil {
-		return "", nil, ui.HandleError(err)
-	}
-
-	prv := []byte(res.PrivateKey)
-	pub = []byte(res.PublicKey)
-	dotSSHPath := getSSHKeyFolder()
-	publicKeyPath := filepath.Join(dotSSHPath, res.Name+".pub")
-	privateKeyPath := filepath.Join(dotSSHPath, res.Name)
-
-	if err = os.WriteFile(privateKeyPath, prv, 0600); err != nil {
-		ui.Errorf("Failed to write private key to %s: %v", privateKeyPath, err)
-		os.Exit(1)
-		return "", nil, nil
-	}
-	if err = os.WriteFile(publicKeyPath, pub, 0600); err != nil {
-		ui.Errorf("Failed to write public key to %s: %v", publicKeyPath, err)
-		os.Exit(1)
-		return "", nil, nil
-	}
-	ui.Successf("Created new SSH key pair:\n"+
-		"  Name: %s\n"+
-		"  Path: %s\n",
-		res.Name, publicKeyPath)
-
-	return res.Name, pub, nil
 }
 
 func sshKeyGenerateFromPrivateKey(ctx context.Context, path string, name *string) (keyName string, pub []byte, err error) {
@@ -200,7 +153,7 @@ func SSHKeyGenerate(cmd *cobra.Command, args []string) error {
 	if len(args) != 0 {
 		name = tools.Stringy(args[0])
 	}
-	_, _, err := sshKeyGenerate(cmd.Context(), config.Config.Unweave.User.ID, name)
+	_, _, err := ssh.Generate(cmd.Context(), config.Config.Unweave.User.ID, name)
 	return err
 }
 
@@ -208,7 +161,7 @@ func SSHKeyList(cmd *cobra.Command, args []string) error {
 	cmd.SilenceUsage = true
 
 	ctx := cmd.Context()
-	uwc := InitUnweaveClient()
+	uwc := config.InitUnweaveClient()
 
 	entries, err := uwc.SSHKey.List(ctx, config.Config.Unweave.User.ID)
 	if err != nil {
