+++ cmd/ssh.go
@@ -1,7 +1,6 @@
 package cmd
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"errors"
@@ -9,151 +8,15 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"strings"
 	"syscall"
 
 	"github.com/spf13/cobra"
 	"github.com/unweave/cli/config"
+	"github.com/unweave/cli/session"
+	"github.com/unweave/cli/ssh"
 	"github.com/unweave/cli/ui"
 	"github.com/unweave/unweave/api/types"
 )
-
-func getUnweaveSSHConfigPath() string {
-	return filepath.Join(config.GetGlobalConfigPath(), "ssh_config")
-}
-
-func addHost(alias, host, user string, port int) error {
-	configEntry := fmt.Sprintf(`Host %s
-    HostName %s
-    User %s
-    Port %d
-    StrictHostKeyChecking no
-    UserKnownHostsFile /dev/null
-    RequestTTY yes
-	ForwardAgent yes
-`, alias, host, user, port)
-
-	file, err := os.OpenFile(getUnweaveSSHConfigPath(), os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0600)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	// Make sure the host block doesn't already exist
-	if e := removeHost(alias); e != nil {
-		ui.Debugf("Failed to remove existing host block: %v", e)
-	}
-
-	if _, err = file.WriteString(configEntry); err != nil {
-		return err
-	}
-
-	home, err := os.UserHomeDir()
-	if err != nil {
-		ui.Errorf("Failed to get user home directory: %v", err)
-		os.Exit(1)
-	}
-	sshConfigPath := filepath.Join(home, ".ssh", "config")
-
-	lines, err := readLines(sshConfigPath)
-	if err != nil {
-		return err
-	}
-
-	// Add to the top of the file if it doesn't already exist
-	includeEntry := "Include " + getUnweaveSSHConfigPath()
-	for _, line := range lines {
-		if strings.HasPrefix(line, includeEntry) {
-			return nil
-		}
-	}
-	lines = append([]string{includeEntry}, lines...)
-
-	return writeLines(sshConfigPath, lines)
-}
-
-func removeHost(alias string) error {
-	lines, err := readLines(getUnweaveSSHConfigPath())
-	if err != nil {
-		return err
-	}
-
-	startIndex := -1
-	for i, line := range lines {
-		if strings.HasPrefix(line, "Host "+alias) {
-			startIndex = i
-			break
-		}
-	}
-
-	if startIndex == -1 {
-		ui.Debugf("Host block not found: %s", alias)
-		return nil
-	}
-	ui.Debugf("Removing host block: %s", alias)
-
-	endIndex := startIndex + 1
-	for i := startIndex + 1; i < len(lines); i++ {
-		if strings.HasPrefix(lines[i], "Host ") {
-			break
-		}
-		endIndex = i
-	}
-
-	// If the host block is not the last block, append the rest of the lines
-	if endIndex != len(lines)-1 {
-		lines = append(lines[:startIndex], lines[endIndex:]...)
-	} else {
-		lines = lines[:startIndex]
-	}
-
-	return writeLines(getUnweaveSSHConfigPath(), lines)
-}
-
-func readLines(path string) ([]string, error) {
-	file, err := os.Open(path)
-	if err != nil {
-		return nil, err
-	}
-	defer file.Close()
-
-	var lines []string
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		lines = append(lines, scanner.Text())
-	}
-	return lines, scanner.Err()
-}
-
-func writeLines(path string, lines []string) error {
-	file, err := os.Create(path)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	writer := bufio.NewWriter(file)
-	for _, line := range lines {
-		fmt.Fprintln(writer, line)
-	}
-	return writer.Flush()
-}
-
-func removeKnownHostsEntry(hostname string) error {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return fmt.Errorf("failed to get user home directory: %v", err)
-	}
-	knownHostsFile := fmt.Sprintf("%s/.ssh/known_hosts", home)
-	removeHostKeyCmd := exec.Command("ssh-keygen", "-R", hostname, "-f", knownHostsFile)
-
-	ui.Debugf("Removing host key from known_hosts: %s", strings.Join(removeHostKeyCmd.Args, " "))
-
-	if err = removeHostKeyCmd.Run(); err != nil {
-		return fmt.Errorf("failed to remove host key from known_hosts: %v", err)
-	}
-	return nil
-}
 
 func copySource(execID, rootDir, dstPath string, connectionInfo types.ConnectionInfo, keyPath string) error {
 	name := fmt.Sprintf("uw-context-%s.tar.gz", execID)
@@ -237,53 +100,6 @@ func copySource(execID, rootDir, dstPath string, connectionInfo types.Connection
 	return nil
 }
 
-func ssh(ctx context.Context, connectionInfo types.ConnectionInfo, args []string) error {
-	overrideUserKnownHostsFile := false
-	overrideStrictHostKeyChecking := false
-
-	for _, arg := range args {
-		if strings.Contains(arg, "UserKnownHostsFile") {
-			overrideUserKnownHostsFile = true
-		}
-		if strings.Contains(arg, "StrictHostKeyChecking") {
-			overrideStrictHostKeyChecking = true
-		}
-	}
-
-	if !overrideUserKnownHostsFile {
-		args = append(args, "-o", "UserKnownHostsFile=/dev/null")
-	}
-	if !overrideStrictHostKeyChecking {
-		args = append(args, "-o", "StrictHostKeyChecking=no")
-	}
-
-	sshCommand := exec.Command(
-		"ssh",
-		append(args, fmt.Sprintf("%s@%s", connectionInfo.User, connectionInfo.Host))...,
-	)
-
-	ui.Debugf("Running SSH command: %s", strings.Join(sshCommand.Args, " "))
-
-	sshCommand.Stdin = os.Stdin
-	sshCommand.Stdout = os.Stdout
-	sshCommand.Stderr = os.Stderr
-
-	if err := sshCommand.Run(); err != nil {
-		if exitError, ok := err.(*exec.ExitError); ok {
-			// Exited with non-zero exit code
-			if status, ok := exitError.Sys().(syscall.WaitStatus); ok {
-				if status.ExitStatus() == 255 {
-					ui.Infof("The remote host closed the connection.")
-					return nil
-				}
-			}
-			return err
-		}
-		return fmt.Errorf("SSH command failed: %v", err)
-	}
-	return nil
-}
-
 func SSH(cmd *cobra.Command, args []string) error {
 	cmd.SilenceUsage = true
 	ctx := cmd.Context()
@@ -353,7 +169,7 @@ func SSH(cmd *cobra.Command, args []string) error {
 			}
 		}
 
-		execCh, errCh, err = execWaitTillReady(watchCtx, execRef)
+		execCh, errCh, err = session.Wait(watchCtx, execRef)
 		if err != nil {
 			return err
 		}
@@ -370,17 +186,17 @@ func SSH(cmd *cobra.Command, args []string) error {
 				}
 				ui.Infof("🚀 Session %q up and running", e.ID)
 
-				if err := removeKnownHostsEntry(e.Connection.Host); err != nil {
+				if err := ssh.RemoveKnownHostsEntry(e.Connection.Host); err != nil {
 					// Log and continue anyway. Most likely the entry is not there.
 					ui.Debugf("Failed to remove known_hosts entry: %v", err)
 				}
 
-				if err := addHost("uw:"+e.ID, e.Connection.Host, e.Connection.User, e.Connection.Port); err != nil {
+				if err := ssh.AddHost("uw:"+e.ID, e.Connection.Host, e.Connection.User, e.Connection.Port); err != nil {
 					ui.Debugf("Failed to add host to ssh config: %v", err)
 				}
 
 				defer func() {
-					if e := removeHost("uw:" + e.ID); e != nil {
+					if e := ssh.RemoveHost("uw:" + e.ID); e != nil {
 						ui.Debugf("Failed to remove host from ssh config: %v", e)
 					}
 				}()
@@ -399,7 +215,7 @@ func SSH(cmd *cobra.Command, args []string) error {
 					ui.Infof("Skipping copying source directory")
 				}
 
-				if err := ssh(ctx, *e.Connection, args); err != nil {
+				if err := ssh.Connect(ctx, *e.Connection, args); err != nil {
 					ui.Errorf("%s", err)
 					os.Exit(1)
 				}
